requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
reportlab==4.0.7
python-magic==0.4.27
aiofiles==23.2.1